Integrated unfinished devgenesis into original one
