+++ scripts/devgenesis.py
@@ -1,10 +1,69 @@
 #!/usr/bin/env python3
 
+import argparse
 import json
 import os
 import re
 import sys
 from pathlib import Path
+from typing import Dict, Any, List
+import subprocess
+import bech32
+
+
+STAKING_DENOM = "afet"
+FUND_BALANCE = 10 ** 23
+
+bonded_pool_address = "fetch1fl48vsnmsdzcv85q5d2q4z5ajdha8yu3xxqtmq"
+not_bonded_pool_address = "fetch1tygms3xhhs3yv487phx3dw4a95jn7t7ljxu6d5"
+VOTING_PERIOD = "60s"
+
+
+
+
+
+def _path(text: str) -> str:
+    return os.path.abspath(text)
+
+
+def _convert_to_valoper(address):
+    hrp, data = bech32.bech32_decode(address)
+    if hrp != 'fetch':
+        print('Invalid address, expected normal fetch address')
+        sys.exit(1)
+
+    return bech32.bech32_encode('fetchvaloper', data)
+
+
+def _from_coin_list(coins: List[Any]) -> Dict[str, int]:
+    balances = {}
+    for coin in coins:
+        balances[str(coin['denom'])] = int(coin['amount'])
+    return balances
+
+
+def _to_coin_list(balances: Dict[str, int]) -> List[Any]:
+    coins = []
+    for denom in sorted(balances.keys()):
+        amount = balances[denom]
+        assert amount >= 0
+
+        if amount == 0:
+            continue
+
+        coins.append({
+            'denom': str(denom),
+            'amount': str(amount)
+        })
+    return coins
+
+def parse_commandline():
+    parser = argparse.ArgumentParser()
+    parser.add_argument('genesis_export', type=_path, help='The path to the genesis export')
+    parser.add_argument('home_path', type=_path, help='The path to the local node data i.e. ~/.fetchd')
+    parser.add_argument('validator_key_name', help='The name of the local key to use for the validator')
+    return parser.parse_args()
+
 
 
 def usage():
@@ -33,72 +92,79 @@ def usage():
     print()
     sys.exit()
 
-
-staking_denom = "afet"
-bonded_pool_address = "fetch1fl48vsnmsdzcv85q5d2q4z5ajdha8yu3xxqtmq"
-not_bonded_pool_address = "fetch1tygms3xhhs3yv487phx3dw4a95jn7t7ljxu6d5"
-voting_period = "60s"
-
-account_to_fund = "fetch1mrf5yyjnnlpy0egvpk2pvjdk9667j2gtu8kpfy"
-fund_balance = 10**23
-
-
 def main():
-    if len(sys.argv) < 2:
+    if len(sys.argv) < 4:
         usage()
 
-    gen_file = sys.argv[1]
-    if not os.path.isfile(gen_file):
-        usage()
+    args = parse_commandline()
 
-    out_homedir = os.path.expanduser(sys.argv[2] if len(sys.argv) > 2 else "~/.fetchd/")
+    print('    Genesis Export:', args.genesis_export)
+    print('  Fetchd Home Path:', args.home_path)
+    print('Validator Key Name:', args.validator_key_name)
 
-    if not os.path.isfile(f"{out_homedir}/config/priv_validator_key.json"):
-        print(f"Cannot find file {out_homedir}/config/priv_validator_key.json")
+    # load up the local validator key
+    local_validator_key_path = os.path.join(args.home_path, 'config', 'priv_validator_key.json')
+    with open(local_validator_key_path, 'r') as input_file:
+        local_validator_key = json.load(input_file)
+
+    # extract the tendermint addresses
+    cmd = ['fetchd', '--home', args.home_path, 'tendermint', 'show-address']
+    validator_address = subprocess.check_output(cmd).decode().strip()
+    validator_pubkey = local_validator_key['pub_key']['value']
+    validator_hexaddr = local_validator_key['address']
+
+    print(f"- new address: {validator_hexaddr}")
+    print(f"- new pubkey: {validator_pubkey}")
+    print(f"- new tendermint address: {validator_address}")
+
+    # extract the address for the local validator key
+    cmd = ['fetchd', '--home', args.home_path, 'keys', 'show', args.validator_key_name, '--output', 'json']
+    key_data = json.loads(subprocess.check_output(cmd).decode())
+
+    if key_data['type'] != 'local':
+        print('Unable to use non-local key type')
         sys.exit(1)
 
-    print(f"Found {out_homedir}/config/priv_validator_key.json")
+    # extract the local address and convert into a valid validator operator address
+    validator_operator_base_address = key_data['address']
+    validator_operator_address = _convert_to_valoper(validator_operator_base_address)
+    print(f'       {validator_operator_base_address}')
+    print(validator_operator_address)
 
-    with open(f"{out_homedir}/config/priv_validator_key.json", "r") as f:
-        priv_validator_key = json.load(f)
 
-    new_hexaddr = priv_validator_key["address"]
-    print(f"- new address: {new_hexaddr}")
-    new_pubkey = priv_validator_key["pub_key"]["value"]
-    print(f"- new pubkey: {new_pubkey}")
-    new_tmaddr = (
-        os.popen(f"fetchd --home {out_homedir} tendermint show-address").read().strip()
-    )
-    print(f"- new tendermint address: {new_tmaddr}")
 
-    with open(gen_file, "r") as f:
-        gen_content = json.load(f)
+    # load the genesis up
+    print('reading genesis export...')
+    with open(args.genesis_export, 'r') as export_file:
+        genesis = json.load(export_file)
+    print('reading genesis export...complete')
 
-    val_infos = gen_content["app_state"]["staking"]["validators"][0]
+
+    val_infos = genesis["app_state"]["staking"]["validators"][0]
     if not val_infos:
         print("Genesis file does not contain any validators")
         sys.exit(1)
 
-    val_opaddr = val_infos["operator_address"]
-    val_pubkey = val_infos["consensus_pubkey"]["key"]
+    target_validator_address = val_infos["operator_address"]
+    target_validator_public_key = val_infos["consensus_pubkey"]["key"]
     val_addr = [
         val
-        for val in gen_content["validators"]
-        if val["pub_key"]["value"] == val_pubkey
+        for val in genesis["validators"]
+        if val["pub_key"]["value"] == target_validator_public_key
     ][0]["address"]
 
     # Replace selected validator by current node one
-    print(f"Replacing validator {val_opaddr}...")
-    gen_content = json.loads(re.sub(val_addr, new_hexaddr, json.dumps(gen_content)))
-    gen_content = json.loads(re.sub(val_pubkey, new_pubkey, json.dumps(gen_content)))
+    print(f"Replacing validator {target_validator_address}...")
+    genesis = json.loads(re.sub(val_addr, validator_hexaddr, json.dumps(genesis)))
+    genesis = json.loads(re.sub(target_validator_public_key, validator_pubkey, json.dumps(genesis)))
 
     # Set .app_state.slashing.signing_infos to contain only our validator signing infos
     print("Updating signing infos...")
-    gen_content["app_state"]["slashing"]["signing_infos"] = [
+    genesis["app_state"]["slashing"]["signing_infos"] = [
         {
-            "address": new_tmaddr,
+            "address": validator_address,
             "validator_signing_info": {
-                "address": new_tmaddr,
+                "address": validator_address,
                 "index_offset": "0",
                 "jailed_until": "1970-01-01T00:00:00Z",
                 "missed_blocks_counter": "0",
@@ -116,95 +182,95 @@ def main():
 
     bonded_tokens = next(
         int(amount["amount"])
-        for balance in gen_content["app_state"]["bank"]["balances"]
+        for balance in genesis["app_state"]["bank"]["balances"]
         if balance["address"] == bonded_pool_address
         for amount in balance["coins"]
-        if amount["denom"] == staking_denom
+        if amount["denom"] == STAKING_DENOM
     )
 
     not_bonded_tokens = next(
         int(amount["amount"])
-        for balance in gen_content["app_state"]["bank"]["balances"]
+        for balance in genesis["app_state"]["bank"]["balances"]
         if balance["address"] == not_bonded_pool_address
         for amount in balance["coins"]
-        if amount["denom"] == staking_denom
+        if amount["denom"] == STAKING_DENOM
     )
 
     not_bonded_tokens = not_bonded_tokens + bonded_tokens - val_tokens
 
-    for balance in gen_content["app_state"]["bank"]["balances"]:
+    for balance in genesis["app_state"]["bank"]["balances"]:
         if balance["address"] == bonded_pool_address:
             for amount in balance["coins"]:
-                if amount["denom"] == staking_denom:
+                if amount["denom"] == STAKING_DENOM:
                     amount["amount"] = str(val_tokens)
 
         if balance["address"] == not_bonded_pool_address:
             for amount in balance["coins"]:
-                if amount["denom"] == staking_denom:
+                if amount["denom"] == STAKING_DENOM:
                     amount["amount"] = str(not_bonded_tokens)
 
     # Create new account and fund it
     print("Creating new account and funding it...")
     # Add new balance to bank
     new_balance = {
-        "address": account_to_fund,
-        "coins": [{"amount": str(fund_balance), "denom": staking_denom}],
+        "address": validator_operator_base_address,
+        "coins": [{"amount": str(FUND_BALANCE), "denom": STAKING_DENOM}],
     }
-    gen_content["app_state"]["bank"]["balances"].append(new_balance)
+    genesis["app_state"]["bank"]["balances"].append(new_balance)
 
     # Add new account to auth
     last_account_number = int(
-        gen_content["app_state"]["auth"]["accounts"][-1]["account_number"]
+        genesis["app_state"]["auth"]["accounts"][-1]["account_number"]
     )
     new_account = {
         "@type": "/cosmos.auth.v1beta1.BaseAccount",
         "account_number": str(last_account_number + 1),
-        "address": account_to_fund,
+        "address": validator_operator_base_address,
         "pub_key": None,
         "sequence": "0",
     }
-    gen_content["app_state"]["auth"]["accounts"].append(new_account)
+    genesis["app_state"]["auth"]["accounts"].append(new_account)
 
     # Update total supply
-    for supply in gen_content["app_state"]["bank"]["supply"]:
-        if supply["denom"] == staking_denom:
-            supply["amount"] = str(int(supply["amount"]) + fund_balance)
+    for supply in genesis["app_state"]["bank"]["supply"]:
+        if supply["denom"] == STAKING_DENOM:
+            supply["amount"] = str(int(supply["amount"]) + FUND_BALANCE)
 
     # Remove all .validators but the one we work with
     print("Removing other validators from initchain...")
-    gen_content["validators"] = [
-        val for val in gen_content["validators"] if val["address"] == new_hexaddr
+    genesis["validators"] = [
+        val for val in genesis["validators"] if val["address"] == validator_hexaddr
     ]
 
     # Set .app_state.staking.last_validator_powers to contain only our validator
     print("Updating last voting power...")
-    gen_content["app_state"]["staking"]["last_validator_powers"] = [
-        {"address": val_opaddr, "power": str(val_power)}
+    genesis["app_state"]["staking"]["last_validator_powers"] = [
+        {"address": target_validator_address, "power": str(val_power)}
     ]
 
     # Jail everyone but our validator
     print("Jail other validators...")
-    for validator in gen_content["app_state"]["staking"]["validators"]:
-        if validator["operator_address"] != val_opaddr:
+    for validator in genesis["app_state"]["staking"]["validators"]:
+        if validator["operator_address"] != target_validator_address:
             validator["status"] = "BOND_STATUS_UNBONDING"
             validator["jailed"] = True
 
-    if "max_wasm_code_size" in gen_content["app_state"]["wasm"]["params"]:
+    if "max_wasm_code_size" in genesis["app_state"]["wasm"]["params"]:
         print("Removing max_wasm_code_size...")
-        del gen_content["app_state"]["wasm"]["params"]["max_wasm_code_size"]
+        del genesis["app_state"]["wasm"]["params"]["max_wasm_code_size"]
 
-    print(f"Setting voting period to {voting_period}...")
-    gen_content["app_state"]["gov"]["voting_params"]["voting_period"] = voting_period
+    print(f"Setting voting period to {VOTING_PERIOD}...")
+    genesis["app_state"]["gov"]["voting_params"]["voting_period"] = VOTING_PERIOD
 
     print("Writing new genesis file...")
-    with open(f"{out_homedir}/config/genesis.json", "w") as f:
-        json.dump(gen_content, f, indent=2)
+    with open(f"{args.home_path}/config/genesis.json", "w") as f:
+        json.dump(genesis, f, indent=2)
 
-    print(f"Done! Wrote new genesis at {out_homedir}/config/genesis.json")
+    print(f"Done! Wrote new genesis at {args.home_path}/config/genesis.json")
     print("You can now start the chain:")
     print()
     print(
-        f"fetchd --home {out_homedir} unsafe-reset-all && fetchd --home {out_homedir} start"
+        f"fetchd --home {args.home_path} unsafe-reset-all && fetchd --home {args.home_path} start"
     )
     print()
